+++ src/utils/error.ts
@@ -12,6 +12,10 @@ export function beautifyStack (fake: Error, real: Error) {
   const stack = fake.stack!.split('\n')
   stack.splice(1, 2)
   stack[0] = real.message
+  /* istanbul ignore if */
+  if (typeof jest === 'object' && real instanceof AssertionError) {
+    stack[0] = ''   // Jest will show error message twice, so we remove it.
+  }
   real.stack = stack.join('\n')
 
   return real