+++ tests/unit/hooks.test.ts
@@ -6,21 +6,19 @@ test('call a function before browser launch', async () => {
   const instance = new Rize({ beforeLaunch: fn })
   await instance
     .execute(() => {
-      expect(fn).toBeCalled()
+      expect(fn).toHaveBeenCalledTimes(1)
     })
     .end()
 }, process.env.CI ? 8000 : 5000)
 
 test('call a function after browser launched', async () => {
-  expect.assertions(4)
-  const fn = jest.fn()
+  const fn = jest.fn(function () {
+    expect(Object.getPrototypeOf(this)).toBe(Rize.prototype)
+  })
   const instance = new Rize({ afterLaunched: fn })
   await instance
-    .execute(function () {
-      expect(this).toBe(instance)
-      expect(this.browser).toBe(instance.browser)
-      expect(this.page).toBe(instance.page)
-      expect(fn).toBeCalled()
+    .execute(() => {
+      expect(fn).toHaveBeenCalledTimes(1)
     })
     .end()
 })
@@ -47,8 +45,10 @@ test('call a function after each step', async () => {
 })
 
 test('call a function before browser exit', async () => {
-  const fn = jest.fn()
+  const fn = jest.fn(function () {
+    expect(Object.getPrototypeOf(this)).toBe(Rize.prototype)
+  })
   const instance = new Rize({ beforeExit: fn })
   await instance.end()
-  expect(fn).toBeCalled()
+  expect(fn).toHaveBeenCalledTimes(1)
 })