+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New API: [`clickLink`](https://rize.js.org/api/classes/_index_.rize.html#clicklink)
 
 ## [0.4.0] - 2018-03-12
 ### Added