+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - New API: [`clickLink`](https://rize.js.org/api/classes/_index_.rize.html#clicklink)
+- New API: [`assertHashIs`](https://rize.js.org/api/classes/_index_.rize.html#asserthashis)
+- New API: [`assertHashBeginsWith`](https://rize.js.org/api/classes/_index_.rize.html#asserthashbeginswith)
 ### Fixed
 - Fix that Jest shows assertion message twice.