+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New API: [`clickLink`](https://rize.js.org/api/classes/_index_.rize.html#clicklink)
 - New API: [`assertHashIs`](https://rize.js.org/api/classes/_index_.rize.html#asserthashis)
 - New API: [`assertHashBeginsWith`](https://rize.js.org/api/classes/_index_.rize.html#asserthashbeginswith)
+- New option: [`browser`](https://rize.js.org/api/interfaces/_index_.rizeoptions.html#browser)
 ### Fixed
 - Fix that Jest shows assertion message twice.