+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New option: [`browser`](https://rize.js.org/api/interfaces/_index_.rizeoptions.html#browser)
 - New lifecycle hook: [`beforeEachStep`](https://rize.js.org/api/interfaces/_index_.rizeoptions.html#beforeeachstep)
 - New lifecycle hook: [`afterEachStep`](https://rize.js.org/api/interfaces/_index_.rizeoptions.html#aftereachstep)
+### Changed
+- Hooks now do not support passing arguments to function
 ### Fixed
 - Fix that Jest shows assertion message twice.
 - Fix missing `this` pointer of `beforeExit` hook.